requirements.txt
Django==3.1.4
psycopg2-binary==2.8.2
django-grappelli
social-auth-app-django==4.0.0
django_compressor
jsonfield==3.1.0
mage2gen
django-jsend
docutils
django-htmlmin
python-memcached
python-social-auth

#API
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0

#Debuging
colorlog
django-debug-toolbar==3.2
django-extensions==3.1.0
Werkzeug==2.0.3

ujson
Pillow==8.0.1
